% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rnet.R
\name{tm_rnet}
\alias{tm_rnet}
\title{Visualise route networks with tmap}
\usage{
tm_rnet(
  rnet,
  palette,
  lwd_multiplier = 4,
  scale = 5,
  lwd = names(rnet)[1],
  col = names(rnet)[1],
  max_features = 10000
)
}
\arguments{
\item{rnet}{}

\item{palette}{}

\item{lwd_multiplier}{The width of the largest line relative to the thinnest line}

\item{scale}{The width of the widest line in the dataset}

\item{lwd}{}

\item{col}{}

\item{max_features}{}
}
\value{
A tmap object
}
\description{
Visualise route networks with tmap
}
\examples{
tm_rnet(rnet, lwd = "ENMAC4")
}
