% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rnet.R
\name{tm_rnet}
\alias{tm_rnet}
\title{Visualise route networks with tmap}
\usage{
tm_rnet(
  rnet,
  palette = "-temperature_diverging",
  lwd_multiplier = 4,
  scale = 5,
  lwd = names(rnet)[1],
  col = names(rnet)[2],
  max_features = 10000,
  breaks = c(0, 25, 50, 75, 100),
  labels = c("0-24", "25-49", "50-75", "75+")
)
}
\arguments{
\item{rnet}{}

\item{palette}{}

\item{lwd_multiplier}{The width of the largest line relative to the thinnest line}

\item{scale}{The width of the widest line in the dataset}

\item{lwd}{}

\item{col}{}

\item{max_features}{}
}
\value{
A tmap object
}
\description{
This function is for visualising route networks.
It was originally developed to show quietness, as defined by
\href{https://www.cyclestreets.net/help/journey/howitworks/}{CycleStreets}
}
\examples{
tm_rnet(rnet_lisbon, lwd = "ENMAC4", col = "Quietness")
tm_rnet(rnet_lisbon, lwd = "ENMAC4", col = "Quietness", scale = 30)
}
